Improve performance of parser with char replace and simpler substring, extract parser into class
